Static methods can be moved to a helper class. A global using will handle the dependancy.
