+++ Restaurant-Menu-Systemv2/GlobalUsings.cs
@@ -1,0 +1,1 @@
+﻿global using static Restaurant_Menu_System_V3.Helpers.OrderHelper;
