+++ Restaurant-Menu-Systemv2/OrderInputAndOptions.cs
@@ -28,66 +28,6 @@
         //set the CustomerReceipt instance enabling rolling total calculation
         public void SetCustomerReceipt(CustomerReceipt receipt) => _customerReceipt = receipt;
 
-        // collects user input and verifies it is valid and within the provided range
-        public static int GetIntegerInput(string message, int minValue, int maxValue)
-        {
-            while (true)
-            {
-                Console.Write(message);
-                if (!int.TryParse(Console.ReadLine(), out int input))
-                {
-                    Console.WriteLine("That is not a number...Please enter one of the provided numbers");
-                }
-                else if (input < minValue || input > maxValue)
-                {
-                    Console.WriteLine("Error! Please enter one of the provided numbers");
-                }
-                else
-                {
-                    return input;
-                }
-            }
-        }
-
-        // displays menu options with their prices
-        public static void DisplayOrderOptions(string orderType, MenuOption[] options)
-        {
-            Console.WriteLine($"\n{orderType}");
-            int itemCount = 1;
-            foreach (MenuOption item in options)
-            {
-                if (item.Price == 0.00m)
-                {
-                    Console.WriteLine($"{itemCount}. {item.ItemName.ToUpper()}");
-                }
-                else
-                {
-                    Console.WriteLine($"{itemCount}. {item.ItemName.ToUpper()}....{item.Price.ToString("c")}");
-                }
-                itemCount++;
-            }
-        }
-
-        // asks a yes/no question and returns the user's response as a boolean
-        public static bool GetYesNoResponse(string question)
-        {
-            while (true)
-            {
-                Console.WriteLine($"{question}(Y/N)");
-                string userResponse = Console.ReadLine() ?? string.Empty;
-
-                if (userResponse.Equals("Y", StringComparison.OrdinalIgnoreCase))
-                {
-                    return true;
-                }
-                else if (userResponse.Equals("N", StringComparison.OrdinalIgnoreCase))
-                {
-                    return false;
-                }
-                Console.WriteLine("Please enter Y/N to continue.");
-            }
-        }
-
         // prompts user to pick their main entrée choice
         public void ChooseEntree()
         {
